Update HolocronMenuComponent.cpp
I completely restructured the holocron logic.
*I gave clear intentions on what I wanted it to do
*Cleaned up the file overall
*Put enough notes so it is easy to understand what each statement is supposed to do
*Added a different cooldown functionality
*Prevented the holocron from being consumed when you're still under cooldown
*Prevented Non jedi (jedi state 2) from using the holocron
*Prevented the holocron from being consumed while full on force

+++ MMOCoreORB/src/server/zone/objects/tangible/components/HolocronMenuComponent.cpp
@@ -3,6 +3,8 @@
  *
  *  Created on: 01/23/2012
  *      Author: xyborn
+ *  Modified on: 10/22/2016
+ *	Author: Skyyyr Death
  */
 
 #include "HolocronMenuComponent.h"
@@ -19,6 +21,20 @@
 
 
 int HolocronMenuComponent::handleObjectMenuSelect(SceneObject* sceneObject, CreatureObject* creature, byte selectedID) const {
+	
+	/*
+		@param If timer is up - Holocron will refil force bar and remove itself
+		@param If timer is down - Holoron will not refil force bar and stay in your inventory
+		
+		If player has !cooldown then
+			refil force bar
+			galactic message (player, "used holocron")
+			player message "your visibility is: #"
+		else
+			player message "your still under cooldown"
+			player message "your visibility is: #"
+	*/
+	
 	if (selectedID != 20)
 		return 0;
 
@@ -27,23 +43,52 @@ int HolocronMenuComponent::handleObjectMenuSelect(SceneObject* sceneObject, Crea
 
 	ManagedReference<PlayerObject*> ghost = creature->getPlayerObject();
 	if (ghost == NULL)
-	return 0;
+		return 0;
 
-	JediManager::instance()->useItem(sceneObject, JediManager::ITEMHOLOCRON, creature);
-        creature->playEffect("clienteffect/pl_force_absorb_hit.cef");
-        sceneObject->destroyObjectFromWorld(true);
-			int jediVis1 = ghost->getVisibility();
-			PlayMusicMessage* pmm = new PlayMusicMessage("sound/music_become_light_jedi.snd");
-  			ghost->sendMessage(pmm);
-			Zone* zone = creature->getZone();
-			//Broadcast to Server
- 			String playerName = creature->getFirstName();
- 			StringBuffer zBroadcast;
- 			zBroadcast << "\\#00E604" << playerName << " \\#63C8F9 Has Used A Holocron";
- 			creature->getZoneServer()->getChatManager()->broadcastGalaxy(NULL, zBroadcast.toString());
-			StringBuffer messageVis;
-			messageVis << "\\#00CC00 Your Visibility is at: " << jediVis1;
-			creature->sendSystemMessage(messageVis.toString());
-
+	//We will check for the holocron cooldown
+	if (!creature->checkCooldownRecovery("used_holocron"))
+		//Message player + visibilty message
+		creature->sendSystemMessage("@jedi_spam:holocron_no_effect");
+		//Visibility message
+		int jediVis1 = ghost->getVisibility();
+		StringBuffer messageVis;
+		messageVis << "\\#00CC00 Your Visibility is at: " << jediVis1;
+		creature->sendSystemMessage(messageVis.toString());
+		return 0;
+	
+	//If you're not a Jedi, don't let them use it.
+	if (playerObject != NULL && playerObject->getJediState() < 2)
+		creature->sendSystemMessage("@jedi_spam:holocron_no_effect");
+		return 0;
+	
+	//full force ? visibility : fillForce
+	if (playerObject->getForcePower() >= playerObject->getForcePowerMax()) {
+		//Message player
+		creature->sendSystemMessage("@jedi_spam:holocron_force_max");
+	} else {
+		//Refil force + Message player
+		creature->sendSystemMessage("@jedi_spam:holocron_force_replenish");
+		playerObject->setForcePower(playerObject->getForcePowerMax(), true);
+		//Set cooldown
+		creature->addCooldown("used_holocron", 1 * 3600000); //3,600,000 = 1 hr
+		//Destroy object
+		sceneObject->destroyObjectFromWorld(true);
+		//Music + Effect
+		creature->playEffect("clienteffect/pl_force_absorb_hit.cef");
+		PlayMusicMessage* pmm = new PlayMusicMessage("sound/music_become_light_jedi.snd");
+  		ghost->sendMessage(pmm);
+		//Broadcast to Server
+		Zone* zone = creature->getZone();
+ 		String playerName = creature->getFirstName();
+ 		StringBuffer zBroadcast;
+ 		zBroadcast << "\\#00E604" << playerName << " \\#63C8F9 Has Used A Holocron";
+		creature->getZoneServer()->getChatManager()->broadcastGalaxy(NULL, zBroadcast.toString());
+	}
+	
+	//After the holocron fills, or doesn't fill your force bar it will tell you your visibility
+	int jediVis1 = ghost->getVisibility();
+	StringBuffer messageVis;
+	messageVis << "\\#00CC00 Your Visibility is at: " << jediVis1;
+	creature->sendSystemMessage(messageVis.toString());
 	return 0;
 }
